changed the tools.py for valuerror issue and made test_classify file
