+++ cellfinder/core/classify/tools.py
@@ -51,13 +51,15 @@ def get_model(
                     "`model_weights` must be provided for inference "
                     "or continued training."
                 )
-
-            raise ValueError(
+            try:
+                model.load_weights(model_weights)
+            except Exception as e:
+                raise ValueError(
                 f"Error loading weights: {model_weights}.\n"
                 "Provided weights don't match the model architecture.\n"
                 "Ensure you are using the correct model file that "
                 "corresponds to these weights."
-            )
+            ) from e
 
     return model
